--- src/components/ui/filter/Filter/index.module.scss
@@ -1,32 +1,0 @@
-@import '../../../../assets/styles/resources';
-
-.filter-component {
-  position: relative;
-  width: 100%;
-  height: 100%;
-
-  > .field {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 56px;
-    height: 56px;
-    background-color: $color-green;
-    border-radius: 50%;
-    box-shadow: $shadow-filter-action;
-
-    > .icon {
-      width: 24px;
-      height: 24px;
-      color: $color-white;
-    }
-  }
-
-  > .panel {
-    position: fixed;
-    top: 0;
-    z-index: 10;
-    width: 100%;
-    height: 100%;
-  }
-}
