--- src/components/ui/filter/Filter/index.ts
@@ -1,1 +1,0 @@
-export * from './Filter';
