+++ src/components/ui/filter/FilterAction/index.module.scss
@@ -1,0 +1,18 @@
+@import '../../../../assets/styles/resources';
+
+.filter-component {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 56px;
+  height: 56px;
+  background-color: $color-green;
+  border-radius: 50%;
+  box-shadow: $shadow-filter-action;
+
+  > .icon {
+    width: 24px;
+    height: 24px;
+    color: $color-white;
+  }
+}
