+++ src/components/ui/filter/FilterAction/index.ts
@@ -1,0 +1,1 @@
+export * from './FilterAction';
